@@ -27,85 +31,89 @@
     "docs": "compodoc -p src/tsconfig.app.json -d docs",
     "docs:serve": "compodoc -p src/tsconfig.app.json -d docs -s",
     "prepush": "npm run lint:ci",
-    "release:changelog": "npm run conventional-changelog -- -p angular -i CHANGELOG.md -s",
-    "dev": "npm run start:hmr",
-    "deploy": "rimraf dist && ng build \"--prod\" \"--aot\" && rm -rf www && cp -r dist www"
+    "release:changelog": "npm run conventional-changelog -- -p angular -i CHANGELOG.md -s"
   },
   "dependencies": {
-    "@agm/core": "1.0.0-beta.0",
-    "@angular/animations": "^4.4.6",
-    "@angular/common": "^4.4.6",
-    "@angular/compiler": "^4.4.6",
-    "@angular/core": "^4.4.6",
-    "@angular/forms": "^4.4.6",
-    "@angular/http": "^4.4.6",
-    "@angular/platform-browser": "^4.4.6",
-    "@angular/platform-browser-dynamic": "^4.4.6",
-    "@angular/router": "^4.4.6",
-    "@asymmetrik/angular2-leaflet": "^2.2.1",
-    "@nebular/auth": "2.0.0-rc.3",
-    "@nebular/theme": "2.0.0-rc.3",
-    "@ng-bootstrap/ng-bootstrap": "1.0.0-beta.5",
-    "@swimlane/ngx-charts": "5.2.0",
-    "angular2-chartjs": "^0.3.0",
-    "angular2-toaster": "^4.0.1",
-    "bootstrap": "4.0.0-beta.2",
-    "chart.js": "^2.7.1",
-    "ckeditor": "^4.8.0",
+    "@agm/core": "1.0.0-beta.2",
+    "@angular/animations": "6.0.0",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/forms": "6.0.0",
+    "@angular/http": "6.0.0",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/router": "6.0.0",
+    "@asymmetrik/ngx-leaflet": "3.0.1",
+    "@nebular/auth": "2.0.0-rc.10",
+    "@nebular/security": "2.0.0-rc.10",
+    "@nebular/theme": "2.0.0-rc.10",
+    "@ng-bootstrap/ng-bootstrap": "1.0.0",
+    "@swimlane/ngx-charts": "7.0.1",
+    "angular-tree-component": "7.2.0",
+    "angular2-chartjs": "0.4.1",
+    "angular2-toaster": "6.0.0",
+    "bootstrap": "4.0.0",
+    "chart.js": "^2.7.3",
+    "ckeditor": "4.7.3",
     "classlist.js": "1.1.20150312",
-    "core-js": "^2.5.3",
-    "d3": "^4.12.2",
-    "font-awesome": "4.7.0",
-    "fundebug-javascript": "^0.3.3",
+    "core-js": "2.5.1",
+    "echarts": "^4.0.2",
+    "fundebug-javascript": "latest",
     "intl": "1.2.5",
     "ionicons": "2.0.1",
-    "leaflet": "^1.2.0",
-    "nebular-icons": "1.0.6",
-    "ng2-ckeditor": "^1.1.13",
-    "ng2-smart-table": "^1.2.2",
-    "ng2-tree": "^2.0.0-alpha.10",
-    "ngx-echarts": "1.2.2",
-    "ngx-webstorage": "1.8.0",
+    "leaflet": "1.2.0",
+    "moment": "^2.22.2",
+    "nebular-icons": "1.0.9",
+    "ng2-adsense": "^5.4.0",
+    "ng2-ckeditor": "1.1.13",
+    "ng2-completer": "2.0.8",
+    "ng2-smart-table": "1.3.5",
+    "ngx-echarts": "2.0.1",
+    "ngx-webstorage": "^2.0.1",
     "normalize.css": "6.0.0",
-    "pace-js": "^1.0.2",
+    "pace-js": "1.0.2",
     "roboto-fontface": "0.8.0",
-    "rxjs": "^5.5.6",
-    "skinview3d": "^1.0.2",
-    "socicon": "^3.0.5",
-    "tether": "^1.4.3",
-    "tinymce": "^4.7.4",
+    "rxjs": "^6.1.0",
+    "rxjs-compat": "^6.1.0",
+    "skinview3d": "^1.1.0",
+    "socicon": "3.0.5",
+    "tinymce": "4.5.7",
     "typeface-exo": "0.0.22",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.14"
+    "web-animations-js": "2.2.5",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6.4",
-    "@angular/compiler-cli": "^4.4.6",
-    "@angular/language-service": "^4.4.6",
-    "@compodoc/compodoc": "^1.0.5",
-    "@types/d3-color": "^1.0.5",
-    "@types/jasmine": "^2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/leaflet": "^1.2.5",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.2.0",
-    "conventional-changelog-cli": "^1.3.5",
+    "@angular-devkit/build-angular": "~0.6.0",
+    "@angular/cli": "6.0.0",
+    "@angular/compiler-cli": "6.0.0",
+    "@angular/language-service": "6.0.0",
+    "@compodoc/compodoc": "^1.1.6",
+    "@fortawesome/fontawesome-free": "^5.2.0",
+    "@types/d3-color": "1.0.5",
+    "@types/googlemaps": "3.30.4",
+    "@types/jasmine": "2.5.54",
+    "@types/jasminewd2": "2.0.3",
+    "@types/leaflet": "1.2.3",
+    "@types/node": "6.0.90",
+    "codelyzer": "4.0.2",
+    "conventional-changelog-cli": "1.3.4",
     "husky": "0.13.3",
-    "jasmine-core": "^2.8.0",
-    "jasmine-spec-reporter": "^4.2.1",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.3.3",
-    "karma-jasmine": "^1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "npm-run-all": "^4.1.2",
-    "protractor": "^5.2.2",
-    "rimraf": "^2.6.2",
+    "jasmine-core": "2.6.4",
+    "jasmine-spec-reporter": "4.1.1",
+    "karma": "1.7.1",
+    "karma-chrome-launcher": "2.1.1",
+    "karma-cli": "1.0.1",
+    "karma-coverage-istanbul-reporter": "1.3.0",
+    "karma-jasmine": "1.1.0",
+    "karma-jasmine-html-reporter": "0.2.2",
+    "npm-run-all": "4.0.2",
+    "protractor": "^5.4.1",
+    "rimraf": "2.6.1",
     "stylelint": "7.13.0",
-    "ts-node": "~3.2.0",
-    "tslint": "^5.9.1",
+    "ts-node": "3.2.2",
+    "tslint": "5.7.0",
     "tslint-language-service": "0.9.6",
-    "typescript": "^2.6.2"
+    "typescript": "2.7.2"
   }
 }
