@@ -2,4 +2,4 @@ export * from './capitalize.pipe';
 export * from './plural.pipe';
 export * from './round.pipe';
 export * from './timing.pipe';
-export * from './rmb.pipe';
+export * from './number-with-commas.pipe';
