@@ -3,29 +3,25 @@ import {NgModule} from '@angular/core';
 import {AuthRoutingModule} from './auth-routing.module';
 import {ThemeModule} from '../@theme/theme.module';
 
-import {NbAuthComponent} from './auth.component';
-import {NbAuthBlockComponent} from './components/auth-block/auth-block.component';
-import {NbLoginComponent} from './components/login/login.component';
-import {NbLogoutComponent} from './components/logout/logout.component';
-import {NbRegisterComponent} from './components/register/register.component';
-import {NbRequestPasswordComponent} from './components/request-password/request-password.component';
-import {NbResetPasswordComponent} from './components/reset-password/reset-password.component';
+import {AuthComponent} from './auth.component';
+import {AuthBlockComponent} from './components/auth-block/auth-block.component';
+import {AuthNoticeComponent} from './components/auth-notice/auth-notice.component';
+import {LoginComponent} from './components/login/login.component';
+import {LogoutComponent} from './components/logout/logout.component';
+import {RegisterComponent} from './components/register/register.component';
+import {RequestPasswordComponent} from './components/request-password/request-password.component';
+import {ResetPasswordComponent} from './components/reset-password/reset-password.component';
 
-import {AuthService} from './services/auth.service';
-import {AuthUtilService} from './services/auth-util.service';
-
-import {UserService} from '../pages/user/user.service';
-import {NeedLoginGuard} from './guards/needLogin.guard';
-import {NeedAdminGuard} from './guards/needAdmin.guard';
 
 const AUTH_COMPONENTS = [
-  NbAuthComponent,
-  NbAuthBlockComponent,
-  NbLoginComponent,
-  NbLogoutComponent,
-  NbRegisterComponent,
-  NbRequestPasswordComponent,
-  NbResetPasswordComponent,
+  AuthComponent,
+  AuthBlockComponent,
+  AuthNoticeComponent,
+  LoginComponent,
+  LogoutComponent,
+  RegisterComponent,
+  RequestPasswordComponent,
+  ResetPasswordComponent,
 ];
 
 @NgModule({
@@ -35,14 +31,6 @@ const AUTH_COMPONENTS = [
   ],
   declarations: [
     ...AUTH_COMPONENTS,
-  ],
-  providers: [
-    UserService,
-    AuthService,
-    AuthUtilService,
-
-    NeedLoginGuard,
-    NeedAdminGuard,
   ],
 })
 
