@@ -1,98 +1,70 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, OnInit, ViewChild} from '@angular/core';
 import {Router} from '@angular/router';
 
-import {AuthService} from '../../services/auth.service';
+import {AuthService} from '../../../@core/data/auth.service';
+import {CommonUtilService} from '../../../@core/utils/common-util.service';
+
+import {AuthNoticeComponent} from '../auth-notice/auth-notice.component';
 
 @Component({
   selector: 'ngx-logout',
   templateUrl: './logout.component.html',
 })
 
-export class NbLogoutComponent implements OnInit {
-  notice: {
-    type: 'info' | 'success' | 'danger',
-    title: string,
-    message: string,
-  };
+export class LogoutComponent implements OnInit {
+  @ViewChild(AuthNoticeComponent) notice: AuthNoticeComponent;
 
   constructor(private router: Router,
-              private authService: AuthService) {
-    this.sendNotice('info', '请稍候', '登出中...');
+              private authService: AuthService,
+              private commonUtilService: CommonUtilService) {
   }
 
-  public ngOnInit(): void {
-    this.authService.getLoginState()
-      .then(userProfile => {
-        if (userProfile['impersonate']) {
-          this.authService.logoutImpersonate()
-            .then(logoutResult => {
-              this.sendNotice('success', '退出替身登陆成功', '即将跳转到dashboard');
+  async ngOnInit(): Promise<void> {
+    this.notice.show(
+      'info',
+      '请稍候',
+      '登出中...',
+    );
 
-              setTimeout(() => {
-                this.router.navigate(['/pages/dashboard'])
-                  .then(navagateState => {
-                    window.location.reload();
-                  });
-              }, 3e3);
-            })
-            .catch(error => {
-              let noticeTitle;
+    try {
+      const loginState = await this.authService.getLoginState();
 
-              switch (error.status) {
-                case 406 : {
-                  noticeTitle = '大兄弟你现在不在替身模式下';
+      if (loginState.impersonate) {
+        await this.authService.logoutImpersonate();
+        this.notice.show(
+          'success',
+          '退出替身登录成功',
+          '即将跳转到dashboard',
+        );
+        await this.commonUtilService.sleep(3e3);
+        await this.router.navigate(['/pages/dashboard']);
+        window.location.reload();
+      } else {
+        await this.authService.logout();
+        this.notice.show(
+          'success',
+          '登出成功',
+          '即将跳转到登录页',
+        );
+        await this.commonUtilService.sleep(3e3);
+        this.router.navigate(['/auth/login']);
+      }
+    } catch (error) {
+      const errorMessageMap = {
+        401: '大兄弟你得先登录',
+        406: '大兄弟你现在不在替身模式下',
+      };
+      const errorMessage = errorMessageMap[error.status] || '未知错误, 请联系鹳狸猿';
 
-                  setTimeout(() => {
-                    this.router.navigate(['/auth/login']);
-                  }, 3e3);
-                  break;
-                }
-                default: {
-                  noticeTitle = '未知错误, 请联系鹳狸猿';
-                }
-              }
+      this.notice.show(
+        'danger',
+        '' + errorMessage,
+        `message: ${error.error.message} | code: ${error.status}`,
+      );
+      console.error(error);
 
-              this.sendNotice('danger', noticeTitle, `message: ${error.error.message || '未知'} | code: ${error.status || '未知'}`);
-            });
-        } else {
-          this.authService.logout()
-            .then(
-              logoutResult => {
-                this.sendNotice('success', '登出成功', '即将跳转到dashboard');
-
-                setTimeout(() => {
-                  this.router.navigate(['/auth/login']);
-                }, 3e3);
-              },
-            )
-            .catch(error => {
-              let noticeTitle;
-
-              switch (error.status) {
-                case 401 : {
-                  noticeTitle = '大兄弟你得先登陆';
-
-                  setTimeout(() => {
-                    this.router.navigate(['/auth/login']);
-                  }, 3e3);
-                  break;
-                }
-                default: {
-                  noticeTitle = '未知错误, 请联系鹳狸猿';
-                }
-              }
-
-              this.sendNotice('danger', noticeTitle, `message: ${error.error.message || '未知'} | code: ${error.status || '未知'}`);
-            });
-        }
-      });
-  }
-
-  private sendNotice(type: 'info' | 'success' | 'danger', title: string, message: string): void {
-    this.notice = {
-      type: type,
-      title: title,
-      message: message,
-    };
+      await this.commonUtilService.sleep(3e3);
+      this.router.navigate(['/auth/login']);
+    }
   }
 }
