@@ -1,7 +1,13 @@
-import {Component, OnInit} from '@angular/core';
-import {ActivatedRoute, Router} from '@angular/router';
+import {Component, OnInit, ViewChild} from '@angular/core';
+import {FormControl, FormGroup, Validators} from '@angular/forms';
+import {Router} from '@angular/router';
 
-import {AuthService} from '../../services/auth.service';
+import {AuthService} from '../../../@core/data/auth.service';
+import {AuthNoticeComponent} from '../auth-notice/auth-notice.component';
+import {CommonUtilService} from '../../../@core/utils/common-util.service';
+import {RouteService} from '../../../@core/services/route.service';
+
+import {passwordEqualValidator} from '../../../@core/directives';
 
 @Component({
   selector: 'ngx-reset-password-page',
@@ -9,66 +15,66 @@ import {AuthService} from '../../services/auth.service';
   templateUrl: './reset-password.component.html',
 })
 
-export class NbResetPasswordComponent implements OnInit {
-  user: any;
-  notice: {
-    type: 'info' | 'success' | 'danger',
-    title: string,
-    message: string,
-  };
-  submitted: boolean;
-  hash: string;
+export class ResetPasswordComponent implements OnInit {
+  public resetPasswordForm: FormGroup;
+  public submitted: boolean;
+  public hash: string;
+  @ViewChild(AuthNoticeComponent) notice: AuthNoticeComponent;
 
   constructor(private router: Router,
-              private activatedRoute: ActivatedRoute,
-              private authService: AuthService) {
-    this.user = {};
+              private authService: AuthService,
+              private commonUtilService: CommonUtilService,
+              private routeService: RouteService) {
     this.submitted = false;
     this.hash = '';
   }
 
-  public ngOnInit(): void {
-    this.activatedRoute.queryParams.subscribe(queryParams => {
-      this.hash = queryParams.hash;
+  public async ngOnInit(): Promise<void> {
+    this.hash = await this.routeService.getQuery('hash');
+
+    this.resetPasswordForm = new FormGroup({
+      password: new FormControl(
+        '', [
+          Validators.required,
+          Validators.minLength(6),
+          Validators.maxLength(32),
+        ],
+      ),
+      repeatPassword: new FormControl(
+        '', [
+          Validators.required,
+          passwordEqualValidator,
+        ],
+      ),
     });
   }
 
-  public resetPassword(): void {
+  public async resetPassword(resetPasswordForm: any): Promise<void> {
     this.submitted = true;
 
-    this.authService.resetPassword(this.hash, this.user.password)
-      .then(
-        resetResult => {
-          this.sendNotice('success', '重置成功', '请使用你的新密码登陆, 即将跳转到登陆页');
+    try {
+      await this.authService.resetPassword(this.hash, resetPasswordForm.password);
+      this.notice.show(
+        'success',
+        '重置成功',
+        '请使用你的新密码登录, 即将跳转到登录页',
+      );
+      await this.commonUtilService.sleep(3e3);
+      this.router.navigate(['/auth/login']);
+    } catch (error) {
+      const errorMessageMap = {
+        403: '令牌无效或已被使用, 请重新找回密码',
+      };
+      const errorMessage = errorMessageMap[error.status] || '未知错误, 请联系鹳狸猿';
 
-          setTimeout(() => {
-            this.router.navigate(['/auth/login']);
-          }, 3e3);
-        },
-      )
-      .catch(error => {
-        let errorTitle;
-        this.submitted = false;
+      this.notice.show(
+        'danger',
+        '' + errorMessage,
+        `message: ${error.error.message} | code: ${error.status}`,
+      );
+      console.error(error);
+    }
 
-        switch (error.status) {
-          case 403 : {
-            errorTitle = '令牌无效或已被使用, 请重新找回密码';
-            break;
-          }
-          default: {
-            errorTitle = '未知错误, 请联系鹳狸猿';
-          }
-        }
-
-        this.sendNotice('danger', errorTitle, `message: ${error.error.message || '未知'} | code: ${error.status || '未知'}`);
-      });
-  }
-
-  private sendNotice(type: 'info' | 'success' | 'danger', title: string, message: string): void {
-    this.notice = {
-      type: type,
-      title: title,
-      message: message,
-    };
+    this.submitted = false;
   }
 }
