@@ -1,7 +1,11 @@
-import {Component, OnInit} from '@angular/core';
-import {ActivatedRoute, Router} from '@angular/router';
+import {Component, OnInit, ViewChild} from '@angular/core';
+import {Router} from '@angular/router';
 
-import {CallbackService} from '../../services/callback.service';
+import {CallbackService} from '../../../@core/data/callback.service';
+import {CommonUtilService} from '../../../@core/utils/common-util.service';
+import {RouteService} from '../../../@core/services/route.service';
+
+import {CallbackNoticeComponent} from '../callback-notice/callback-notice.component';
 
 @Component({
   selector: 'ngx-oauth-qq',
@@ -9,38 +13,40 @@ import {CallbackService} from '../../services/callback.service';
 })
 
 export class OauthQQComponent implements OnInit {
-  notice: {
-    type: 'info' | 'success' | 'danger',
-    title: string,
-    message: string,
-  };
+  @ViewChild(CallbackNoticeComponent) notice: CallbackNoticeComponent;
 
   constructor(private router: Router,
-              private activatedRoute: ActivatedRoute,
-              private callbackService: CallbackService) {
-    this.sendNotice('info', '请稍后...', '授权中...');
+              private callbackService: CallbackService,
+              private commonUtilService: CommonUtilService,
+              private routeService: RouteService) {
   }
 
-  public ngOnInit(): void {
-    this.activatedRoute.queryParams.subscribe(queryParams => {
-      this.callbackService.oauthQQCallback(queryParams.accessToken, queryParams.expiresIn)
-        .then(success => {
-          this.sendNotice('success', '授权成功', '即将跳转到dashboard');
-          setTimeout(() => {
-            this.router.navigate(['/pages/user/socials']);
-          }, 3e3);
-        })
-        .catch(error => {
-          this.sendNotice('danger', '授权失败', `message: ${error.error.message || '未知'} | code: ${error.status || '未知'}`);
-        });
-    });
-  }
+  public async ngOnInit(): Promise<void> {
+    this.notice.show(
+      'info',
+      '请稍候...',
+      '授权中...',
+    );
 
-  private sendNotice(type: 'info' | 'success' | 'danger', title: string, message: string): void {
-    this.notice = {
-      type: type,
-      title: title,
-      message: message,
-    };
+    try {
+      const accessToken = await this.routeService.getQuery('accessToken');
+      const expiresIn = await this.routeService.getQuery('expiresIn');
+
+      await this.callbackService.oauthQQCallback(accessToken, expiresIn);
+      this.notice.show(
+        'success',
+        '授权成功',
+        '即将跳转到社交账户授权',
+      );
+      await this.commonUtilService.sleep(3e3);
+      this.router.navigate(['/pages/user/socials']);
+    } catch (error) {
+      this.notice.show(
+        'danger',
+        '授权登录失败',
+        `message: ${error.error.message || '未知'} | code: ${error.status || '未知'}`,
+      );
+      console.error(error);
+    }
   }
 }
