@@ -2,13 +2,13 @@ import {NgModule} from '@angular/core';
 
 import {ThemeModule} from '../../../@theme/theme.module';
 import {UsersModule} from './users/users.module';
-import {UserModule} from './user/user.module';
+import {UserModule} from './users/user/user.module';
 
 import {RmcaRoutingModule} from './rmca-routing.module';
 
 import {RmcaComponent} from './rmca.component';
 import {UsersComponent} from './users/users.component';
-import {UserComponent} from './user/user.component';
+import {UserComponent} from './users/user/user.component';
 
 import {RmcaService} from './rmca.service';
 
