@@ -2,12 +2,14 @@ import {Component, OnInit} from '@angular/core';
 import {Router} from '@angular/router';
 import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
 
-import {NoticeService} from '../../../../@system/notice/notice.service';
 import {RmcaService} from '../rmca.service';
 
-import {AdminRmcaUsersUser} from '../../../../@model/admin/rmca/users/user.interface';
-import {UserBanModalComponent} from './user-ban-modal/user-ban-modal.component';
-import {UserUnbanModalComponent} from './user-unban-modal/user-unban-modal.component';
+import {IAdminRmcaUsersUser} from '../../../../@model/common/admin/rmca/users/user.interface';
+import {UserBanModalComponent} from './user/user-ban-modal/user-ban-modal.component';
+import {UserUnbanModalComponent} from './user/user-unban-modal/user-unban-modal.component';
+import {AuthUtilService} from '../../../../@core/utils/auth-util.service';
+import {IUser, UserState} from '../../../../@model/common/user/user.interface';
+import {NoticeUtilService} from '../../../../@core/utils/notice-util.service';
 
 @Component({
   styleUrls: ['./users.component.scss'],
@@ -15,88 +17,72 @@ import {UserUnbanModalComponent} from './user-unban-modal/user-unban-modal.compo
 })
 
 export class UsersComponent implements OnInit {
-  filter: {
-    type: string,
-    types: any,
-  };
-  search: {
-    keyWord: string,
-    submitted: boolean;
-  };
-  users: AdminRmcaUsersUser[];
-  page: number;
-  limit: number;
+  public users: IAdminRmcaUsersUser[];
+  public UserStateEnum = UserState;
 
-  constructor(private noticeService: NoticeService,
+  public page: number;
+  public limit: number;
+  public count: number;
+
+
+  public loading: boolean;
+
+  constructor(private noticeUtilService: NoticeUtilService,
               private modalService: NgbModal,
               private rmcaService: RmcaService,
-              private router: Router) {
-    this.search = {
-      keyWord: '',
-      submitted: false,
-    };
-    this.filter = {
-      type: 'all',
-      types: {
-        all: 'all',
-        onlyAdmin: 'onlyAdmin',
-        onlyBan: 'onlyBan',
-      },
-    };
+              private router: Router,
+              private authUtilService: AuthUtilService) {
     this.users = [];
     this.page = 1;
     this.limit = 10;
+    this.count = 1;
+    this.loading = false;
   }
 
   public ngOnInit(): void {
     this.getUsers(this.page, this.limit);
   }
 
-  public filterAll(): void {
-    this.page = 1;
-    this.getUsers(this.page, this.limit);
-  }
+  public async getUsers(page: number, limit: number): Promise<void> {
+    this.loading = true;
 
-  public filterAdmin(): void {
-    this.page = 1;
-    this.searchUser(this.search.keyWord, this.page, this.limit, true, false);
-  }
-
-  public filterBanned(): void {
-    this.page = 1;
-    this.searchUser(this.search.keyWord, this.page, this.limit, false, true);
-  }
-
-  public searchClick(): void {
-    this.page = 1;
-    this.searchUser(this.search.keyWord, this.page, this.limit, this.filter.type == this.filter.types.onlyAdmin, this.filter.type == this.filter.types.onlyBan);
-  }
-
-  public searchKeyDown(event): void {
-    if (event.keyCode == 13) {
-      this.searchClick();
+    try {
+      const usersResponse = await this.rmcaService.getUsers(page, limit);
+      const users = usersResponse['data'];
+      this.count = usersResponse['count'];
+      this.users = users.map((user: IUser) => this.authUtilService.extendUserModel(user));
+    } catch (error) {
+      this.noticeUtilService.errorNotice(error, '获取用户列表失败');
     }
+
+    this.loading = false;
   }
 
-  public pageAdd(): void {
-    this.page++;
-    this._pageChange(this.page);
-  }
 
-  public pageKeyDown(event): void {
-    if (event.keyCode == 13) {
-      this._pageChange(this.page);
-    }
-  }
+  // public filterAll(): void {
+  //   this.page = 1;
+  //   this.getUsers(this.page, this.limit);
+  // }
+  //
+  // public filterAdmin(): void {
+  //   this.page = 1;
+  //   this.searchUser(this.search.keyWord, this.page, this.limit, true, false);
+  // }
+  //
+  // public filterBanned(): void {
+  //   this.page = 1;
+  //   this.searchUser(this.search.keyWord, this.page, this.limit, false, true);
+  // }
+  //
+  // public searchClick(): void {
+  //   this.page = 1;
+  //   this.searchUser(this.search.keyWord, this.page, this.limit,
+  //     this.filter.type === this.filter.types.onlyAdmin,
+  //     this.filter.type === this.filter.types.onlyBan);
+  // }
 
-  public pageMinus(): void {
-    if (this.page > 1) {
-      this.page--;
-      this._pageChange(this.page);
-    }
-  }
 
-  public banUser(user: AdminRmcaUsersUser): void {
+  public openBanUserModal(user: IAdminRmcaUsersUser): void {
     const activeModal = this.modalService.open(UserBanModalComponent, {
       size: 'lg',
       container: 'nb-layout',
@@ -109,7 +95,7 @@ export class UsersComponent implements OnInit {
     });
   }
 
-  public unBanUser(user: AdminRmcaUsersUser): void {
+  public openUnBanUserModal(user: IAdminRmcaUsersUser): void {
     const activeModal = this.modalService.open(UserUnbanModalComponent, {
       size: 'lg',
       container: 'nb-layout',
@@ -122,36 +108,24 @@ export class UsersComponent implements OnInit {
     });
   }
 
-  private getUsers(page: number, limit: number): void {
-    this.rmcaService.getUsers(page, limit)
-      .then(users => {
-        this.users = users as AdminRmcaUsersUser[];
-      })
-      .catch(error => {
-        this.noticeService.error('获取用户列表失败, 请刷新页面重试', `message: ${error.error.message || '未知'} | code: ${error.status || '未知'}`);
-      });
-  }
-
   public goToUserDetail(userId): void {
     this.router.navigate(['/pages/admin/rmca/user', userId]);
   }
 
-  private searchUser(keyword: string, page: number, limit: number, isAdmin?: boolean, isBanned?: boolean): void {
-    this.search.submitted = true;
-
-    this.rmcaService.searchUsers(keyword, page, limit, isAdmin, isBanned)
-      .then(users => {
-        this.search.submitted = false;
-        this.users = users as AdminRmcaUsersUser[];
-      })
-      .catch(error => {
-        this.search.submitted = false;
-        this.noticeService.error('获取搜索结果失败', `message: ${error.error.message || '未知'} | code: ${error.status || '未知'}`);
-      });
-  }
-
-  private _pageChange(page): void {
-    this.filter.type = this.filter.types.all;
-    this.getUsers(page, this.limit);
-  }
+  // private searchUser(keyword: string, page: number, limit: number, isAdmin?: boolean, isBanned?: boolean): void {
+  //   this.search.submitted = true;
+  //
+  //   this.rmcaService.searchUsers(keyword, page, limit, isAdmin, isBanned)
+  //     .then(users => {
+  //       this.search.submitted = false;
+  //       this.users = users as IAdminRmcaUsersUser[];
+  //     })
+  //     .catch(error => {
+  //       this.search.submitted = false;
+  //       const errorMessageMap = {};
+  //       const errorMessage = errorMessageMap[error.status] || `[${error.status}] ${error.error.message}`;
+  //       this.noticeService.error('获取搜索结果失败', errorMessage);
+  //       console.error(error);
+  //     });
+  // }
 }
