@@ -2,8 +2,8 @@ import {NgModule} from '@angular/core';
 
 import {ThemeModule} from '../../../../@theme/theme.module';
 
-import {UserBanModalComponent} from './user-ban-modal/user-ban-modal.component';
-import {UserUnbanModalComponent} from './user-unban-modal/user-unban-modal.component';
+import {UserBanModalComponent} from './user/user-ban-modal/user-ban-modal.component';
+import {UserUnbanModalComponent} from './user/user-unban-modal/user-unban-modal.component';
 
 const components = [
   UserBanModalComponent,
