@@ -6,18 +6,14 @@ import {PlayerRoutingModule, routedComponents} from './player-routing.module';
 
 import {PlayerComponent} from './player.component';
 import {RolesComponent} from './roles/roles.component';
-import {SkinViewerComponent} from './roles/skin-viewer/skin-viewer.component';
 import {RoleAddModalComponent} from './roles/role-add-modal/role-add-modal.component';
 import {RoleDetailModalComponent} from './roles/role-detail-modal/role-detail-modal.component';
 
-import {PlayerService} from './player.service';
-import {UserService} from '../user/user.service';
-import {NoticeService} from '../../@system/notice/notice.service';
+import {PlayerService} from '../../@core/data/player.service';
 
 const components = [
   PlayerComponent,
   RolesComponent,
-  SkinViewerComponent,
   RoleAddModalComponent,
   RoleDetailModalComponent,
 ];
@@ -33,8 +29,6 @@ const components = [
   ],
   providers: [
     PlayerService,
-    UserService,
-    NoticeService,
   ],
   entryComponents: [
     RoleAddModalComponent,
