@@ -1,57 +1,59 @@
-import {Component, EventEmitter, Output} from '@angular/core';
+import {Component, EventEmitter, OnInit, Output} from '@angular/core';
+import {FormControl, FormGroup, Validators} from '@angular/forms';
 import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
 
-import {PlayerService} from '../../player.service';
-import {NoticeService} from '../../../../@system/notice/notice.service';
+import {PlayerService} from '../../../../@core/data/player.service';
+import {NoticeService} from '../../../../@core/services/notice.service';
+import {NoticeUtilService} from '../../../../@core/utils/notice-util.service';
 
 @Component({
   styleUrls: ['./role-add-modal.component.scss'],
   templateUrl: './role-add-modal.component.html',
 })
 
-export class RoleAddModalComponent {
+export class RoleAddModalComponent implements OnInit {
   @Output() event = new EventEmitter();
-  submitted: boolean;
 
-  constructor(private playerService: PlayerService,
+  public roleForm: FormGroup;
+  public submitted: boolean;
+
+  constructor(private noticeUtilService: NoticeUtilService,
               private noticeService: NoticeService,
-              private activeModal: NgbActiveModal) {
+              private playerService: PlayerService,
+              public activeModal: NgbActiveModal) {
     this.submitted = false;
   }
 
-  public addRole(roleForm): void {
-    this.submitted = true;
-
-    this.playerService.addRole(roleForm.rolename)
-      .then(createState => {
-        this.noticeService.success('新增成功', '新增角色成功');
-        this.event.emit();
-        this.activeModal.close();
-      })
-      .catch(error => {
-        this.submitted = false;
-
-        let errorMessage = '';
-
-        switch (error.status) {
-          case 409: {
-            errorMessage = '角色名已存在';
-            break;
-          }
-          case 450: {
-            errorMessage = '当前角色数量超过用户账号限制';
-            break;
-          }
-          default: {
-            errorMessage = `message: ${error.error.message || '未知'} | code: ${error.status || '未知'}`;
-          }
-        }
-
-        this.noticeService.error('新增角色失败', errorMessage);
-      });
+  public ngOnInit(): void {
+    this.roleForm = new FormGroup({
+      rolename: new FormControl(
+        '', [
+          Validators.required,
+          Validators.pattern('^[a-zA-Z0-9_-]{1,16}$'),
+        ],
+      ),
+    });
   }
 
-  public closeModal(): void {
-    this.activeModal.close();
+  public async addRole(roleForm: any): Promise<void> {
+    this.submitted = true;
+
+    try {
+      await this.playerService.addRole(roleForm.rolename);
+      this.noticeService.success(
+        '新增成功',
+        '新增角色成功',
+      );
+      this.event.emit(roleForm);
+      this.activeModal.close();
+    } catch (error) {
+      const errorMessageMap = {
+        409: '角色名已存在',
+        450: '当前角色数量超过用户账号限制',
+      };
+      this.noticeUtilService.errorNotice(error, '新增角色失败', errorMessageMap);
+    }
+
+    this.submitted = false;
   }
 }
