@@ -1,13 +1,18 @@
 import {Component, OnInit} from '@angular/core';
 
-import {NoticeService} from '../../../@system/notice/notice.service';
+import {IUserExtend, UserState} from '../../../@model/common/user/user.interface';
+import {AuthUtilService} from '../../../@core/utils/auth-util.service';
+import {NoticeUtilService} from '../../../@core/utils/notice-util.service';
 
-import {UserService} from '../user.service';
+export interface IUserExtendProfile extends IUserExtend {
+  stateColor: StateColor;
+}
 
-import {User} from '../../../@model/user/user.interface';
-import {DefaultUser} from '../../../@model/user/user.const';
-import {UserCacheService} from '../../../@system/cache/service/user-cache.service';
-
+export enum StateColor {
+  danger = 'danger',
+  warning = 'warning',
+  success = 'success',
+}
 
 @Component({
   styleUrls: ['./profile.component.scss'],
@@ -15,27 +20,39 @@ import {UserCacheService} from '../../../@system/cache/service/user-cache.servic
 })
 
 export class ProfileComponent implements OnInit {
-  user: User;
+  user: IUserExtendProfile;
 
-  constructor(private noticeService: NoticeService,
-              private userService: UserService,
-              private userCacheService: UserCacheService) {
-    this.user = DefaultUser;
+  constructor(private noticeUtilService: NoticeUtilService,
+              private authUtilService: AuthUtilService) {
   }
 
   public ngOnInit(): void {
-    this.user = this.userCacheService.getCache();
+    this.user = ProfileComponent.extendUser(this.authUtilService.user);
   }
 
-  public async getUserProfile(): Promise<void> {
+  public async updateUserProfile(): Promise<void> {
     try {
-      const user = await this.userService.getUserProfile();
-
-      this.userCacheService.setCache(user as User);
-      this.user = user as User;
+      this.user = ProfileComponent.extendUser(await this.authUtilService.updateUser());
     } catch (error) {
-      this.noticeService.error('获取用户信息失败, 请刷新页面重试', `message: ${error.error.message || '未知'} | code: ${error.status || '未知'}`);
-      console.trace(error);
+      this.noticeUtilService.errorNotice(error, '获取用户信息失败');
     }
+  }
+
+  private static extendUser(user: IUserExtend): IUserExtendProfile {
+    const extendUserProfile = user as IUserExtendProfile;
+
+    switch (user.state) {
+      case UserState.BANNED:
+        extendUserProfile.stateColor = StateColor.danger;
+        break;
+      case UserState.NEED_EMAIL_VALIDATION:
+        extendUserProfile.stateColor = StateColor.warning;
+        break;
+      case UserState.NORMAL:
+      default:
+        extendUserProfile.stateColor = StateColor.success;
+    }
+
+    return extendUserProfile;
   }
 }
