@@ -6,27 +6,22 @@ import {routedComponents, UserRoutingModule} from './user-routing.module';
 import {UserComponent} from './user.component';
 import {ProfileComponent} from './profile/profile.component';
 import {ProfileUserInfoComponent} from './profile/user-info/profile-user-info.component';
-import {ProfileUserPasswordComponent} from './profile/user-password/profile-user-password.component';
-import {ProfileUserStateComponent} from './profile/user-state/profile-user-state.component';
-import {YggdrasilComponent} from './yggdrasil/yggdrasil.component';
-import {YggdrasilInfoComponent} from './yggdrasil/yggdrasil-info/yggdrasil-info.component';
-import {YggdrasilStateComponent} from './yggdrasil/yggdrasil-state/yggdrasil-state.component';
+import {ProfileYggdrasilInfoComponent} from './profile/yggdrasil-info/profile-yggdrasil-info.component';
+import {ProfileUpdatePasswordComponent} from './profile/update-password/profile-update-password.component';
+
 import {SocialsComponent} from './socials/socials.component';
 import {SocialGravatarStateComponent} from './socials/gravatar-state/social-gravatar-state.component';
 import {SocialQQStateComponent} from './socials/qq-state/social-QQ-state.component';
 import {SocialWeiboStateComponent} from './socials/weibo-state/social-weibo-state.component';
 
-import {UserService} from './user.service';
-
 const components = [
   UserComponent,
+
   ProfileComponent,
   ProfileUserInfoComponent,
-  ProfileUserPasswordComponent,
-  ProfileUserStateComponent,
-  YggdrasilComponent,
-  YggdrasilInfoComponent,
-  YggdrasilStateComponent,
+  ProfileYggdrasilInfoComponent,
+  ProfileUpdatePasswordComponent,
+
   SocialsComponent,
   SocialGravatarStateComponent,
   SocialQQStateComponent,
@@ -41,9 +36,6 @@ const components = [
   declarations: [
     routedComponents,
     ...components,
-  ],
-  providers: [
-    UserService,
   ],
 })
 
